pyproject.toml
[project]
name = "dsgp-15-project"
version = "0.1.0"
description = "Deep learning project to analyze children's drawings for emotion recognition and creativity assessment."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "albumentations>=2.0.8",
    "chromadb>=1.4.1",
    "contourpy>=1.3.3",
    "fastapi>=0.122.0",
    "ipykernel>=7.1.0",
    "jupyterlab>=4.5.0",
    "kagglehub>=0.3.13",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-text-splitters>=1.1.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.2.2",
    "shapely>=2.1.2",
    "tensorflow>=2.20.0",
    "tf-keras>=2.20.1",
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "torchaudio>=2.2.0",
    "tqdm>=4.67.1",
    "transformers[torch]>=4.40.0",
    "ultralytics>=8.1,<9.0",
    "uvicorn>=0.38.0",
    "sqlalchemy>=2.0.48",
    "segment-anything>=1.0",
    "streamlit>=1.55.0",
]
